Loggers
Separate out loggers into their own directory and files
