+++ src/services/discord/loggers/memberRemoveLogger.js
@@ -1,0 +1,12 @@
+const { MessageEmbed } = require('discord.js')
+const { main_server, colors } = require('../info')
+module.exports = (member) => {
+  if (member.guild.id !== main_server) return
+  const embed = new MessageEmbed()
+    .setAuthor(member.user.tag, member.user.avatarURL())
+    .setColor(colors['blue'])
+    .setDescription(`<@` + member.id + `> has left.`)
+    .setFooter('User ID: ' + member.id)
+  channel['logs'].send(embed)
+  channel['member_count'].setName('Alpine Fam: ' + member.guild.memberCount)
+}
